==> 0-0-get-raw-data2input.ipynb <==
import os, collections
import numpy as np
import pandas as pd

RAW_PATH = "/data2/zhoujb/project/cowpea_project/rawData/"

#处理原始的SNP文件，分成每一个染色体一个文件
raw_snp = pd.read_table(os.path.join(RAW_PATH, "AK808.ho.0.1.snplist"))
for group in raw_snp.groupby("#Chr"):
    group_name, group_data = group
    if group_name in ['Chr01', 'Chr02', 'Chr03', 'Chr04', 'Chr05', 'Chr06', 'Chr07', 'Chr08', 'Chr09', 'Chr10', 'Chr11']:
        group_data.to_csv(os.path.join(RAW_PATH, "{}.snplist".format(group_name)), sep="\t", index=0)

raw_snp["#Chr"].value_counts()

raw_region = pd.read_excel(os.path.join(RAW_PATH, "XP-CLR.xlsx"))
raw_region = raw_region.drop_duplicates(subset=['Chromosome', 'Region start (bp)', 'Region end (bp)'])
raw_region = raw_region.sort_values(by=['Chromosome', 'Region start (bp)', 'Region end (bp)'])
raw_region.to_csv(os.path.join(RAW_PATH, "XP-CLR_region.txt"), sep="\t", index=0)

raw_region[raw_region["Region start (bp)"]==1]

detail_info = pd.read_excel(os.path.join(RAW_PATH, "Detail_information_344.xlsx"), index_col="Genotype No.")
detail_info["Type"] = detail_info["Types"].map({'Grain':"G", 'V-landrace':"VL", '-':"NA", 'V-cultivar':"VC"})

detail_info["Type"].value_counts()
