==> 1-1-getSNPCluster.ipynb <==
import os
import numpy as np
import pandas as pd

from mpire import WorkerPool

RAW_PATH = "/data2/zhoujb/project/cowpea_project/rawData/"
SNP_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/snpDir/"
CLUSTER_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/SNPMMSeqCluster/"

def SNP2Fasta(input_file):
    raw_data = pd.read_table(os.path.join(SNP_PATH, input_file))
    sample_list = list(raw_data.columns[2:])
    #sample_list = list(raw_data.columns[3:])

    with open(os.path.join(CLUSTER_PATH, '{}.fa'.format(input_file.split(".")[0])), "w") as out_f:
        for sample_name in sample_list:
            print(">{}".format(sample_name), file=out_f)
            print("".join(raw_data[sample_name].values), file=out_f)

    return

if __name__ == "__main__":
    in_list = [x for x in os.listdir(SNP_PATH) if x.endswith("txt")]
    with WorkerPool(n_jobs=48) as pool:
        pool.map(SNP2Fasta, in_list, progress_bar=True)

print("ALL DONE")
