==> 1-2-run-mmseq.ipynb <==
import os, subprocess
import numpy as np
import pandas as pd

from mpire import WorkerPool

CLUSTER_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/SNPMMSeqCluster/"

def runMMSeq(file_name):
    cmd = "/data2/zhoujb/anaconda3/envs/cropgbm21/bin/mmseqs easy-cluster {} {} tmp_{} --min-seq-id 0.9 --cov-mode 0 --alignment-mode 3 --similarity-type 2 --cluster-mode 1 --threads 4".format(file_name, file_name.split(".")[0], file_name.split(".")[0])
    subprocess.run(cmd, shell=True)

    return
    
if __name__ == "__main__":
    file_list = [x for x in os.listdir(CLUSTER_PATH) if x.endswith("fa")]
    with WorkerPool(n_jobs=48) as pool:
        pool.map(runMMSeq, file_list, progress_bar=True)

print("ALL DONE")
