==> 1-4-rm-cluster.ipynb <==
import os
import numpy as np
import pandas as pd

from mpire import WorkerPool

RAW_PATH = "/data2/zhoujb/project/cowpea_project/rawData/"
SNP_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/snpDir/"
CLUSTER_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/SNPMMSeqCluster/"

def getFinalCluser(input_file):
    input_df = pd.read_table(os.path.join(CLUSTER_PATH, input_file), header=None, names=["Class_Rep", "Sample_name"])
    input_count = input_df.groupby("Class_Rep").count()
    input_count["Sample_pertange"] = input_count["Sample_name"] / len(input_df)
    max_pertange = input_count["Sample_pertange"].max()
    if max_pertange > 0.5:
        input_df.to_csv(os.path.join(CLUSTER_PATH, "{}.final".format(input_file.split(".")[0])), sep="\t", index=0)
    return

if __name__ == "__main__":
    file_list = [x for x in os.listdir(CLUSTER_PATH) if x.endswith("_cluster.tsv")]
    with WorkerPool(n_jobs=48) as pool:
        pool.map(getFinalCluser, file_list, progress_bar=True)

print("ALL DONE")
