==> 3-0-get-score.ipynb <==
import os, pickle
import numpy as np
import pandas as pd

RAW_PATH = "/data2/zhoujb/project/cowpea_project/rawData/"
SNP_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/snpDir/"
CLUSTER_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/SNPMMSeqCluster/"
RES_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/Result/"

haplotype_classtype_mean = pd.DataFrame(columns=["Haplotype_Class", "Mean.Total_freq", "Mean.VL_freq", "Mean.G_freq", "Mean.VC_freq"])
haplotype_counts_info = pd.read_table(os.path.join(RES_PATH, "Haplotype_Counts_Info.tsv"))
for group in haplotype_counts_info.groupby("Haplotype_class"):
    group_name, group_data = group
    haplotype_classtype_mean.loc[len(haplotype_classtype_mean)] = [group_name, group_data["Genotyped_prop"].mean(), 
                                                                   group_data["VL_prop"].mean(), 
                                                                   group_data["G_prop"].mean(), 
                                                                   group_data["VC_prop"].mean()]

haplotype_classtype_mean = haplotype_classtype_mean.set_index("Haplotype_Class")
haplotype_classtype_mean.to_csv(os.path.join(RES_PATH, "Haplotype_ClassType_Mean.Freq.tsv"), sep="\t")

haplotype_classtype_mean

gene_num_dict = {}
raw_region_info = pd.read_table(os.path.join(RAW_PATH, "XP-CLR_region.txt"))
gene_num_median = raw_region_info["Gene Number"].median()
for index in raw_region_info.index:
    Chr = raw_region_info.loc[index, 'Chromosome']
    Start = raw_region_info.loc[index, 'Region start (bp)']
    End = raw_region_info.loc[index, 'Region end (bp)']
    gene_num = raw_region_info.loc[index, 'Gene Number']
    gene_num_dict["{}_{}_{}".format(Chr, Start, End)] = gene_num / gene_num_median

gene_num_dict

len(gene_num_dict)

haplotype_score_info = pd.DataFrame(columns=["Haplotype_ID", "Haplotype_Class", "Gene_Num_Adj_Coeff", 
                                             "Mean.Total_freq", "Mean.VL_freq","Mean.G_freq", "Mean.VC_freq", 
                                             "Hap_Score.by_Total_pop", "Hap_Score.by_VL_pop", "Hap_Score.by_G_pop", "Hap_Score.by_VC_pop"])
for index in haplotype_counts_info.index:
    haplotype_id = 	haplotype_counts_info.loc[index, "Haplotype_id"]
    Chr = haplotype_counts_info.loc[index, "Chr"]
    Start = haplotype_counts_info.loc[index, "Start"]
    End = haplotype_counts_info.loc[index, "End"]
    haplotype_class = haplotype_counts_info.loc[index, "Haplotype_class"]
    haplotype_mean = haplotype_classtype_mean.loc[haplotype_class].values
    haplotype_score = haplotype_mean * gene_num_dict["{}_{}_{}".format(Chr, Start, End)]
    haplotype_score_info.loc[len(haplotype_score_info)] = [haplotype_id, haplotype_class, gene_num_dict["{}_{}_{}".format(Chr, Start, End)]] + list(haplotype_mean) + list(haplotype_score)

haplotype_score_info.to_csv(os.path.join(RES_PATH, "Haplotype_Score_Info.tsv"), sep="\t", index=0)

haplotype_score_info
