==> 3-1-get-total-best-hap.ipynb <==
import os, pickle
import numpy as np
import pandas as pd

RAW_PATH = "/data2/zhoujb/project/cowpea_project/rawData/"
SNP_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/snpDir/"
CLUSTER_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/SNPMMSeqCluster/"
RES_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/Result/"

total_best_haplotype_comb = pd.DataFrame()
haplotype_counts_info = pd.read_table(os.path.join(RES_PATH, "Haplotype_Counts_Info.tsv"))
haplotype_counts_info = haplotype_counts_info[['Haplotype_id', 'Haplotype_class', 'Chr', 'Start', 'End', 
                                               'Genotyped_prop', 'VL_prop', 'G_prop', 'VC_prop']].copy()
for group in haplotype_counts_info.groupby(['Chr', 'Start', 'End']):
    group_name, group_data = group
    tmp_data = group_data[group_data["Genotyped_prop"] == group_data["Genotyped_prop"].max()]
    total_best_haplotype_comb = pd.concat([total_best_haplotype_comb, tmp_data], axis=0, sort=False, ignore_index=True)

total_best_haplotype_comb = total_best_haplotype_comb.set_index("Haplotype_id")

haplotype_score_info = pd.read_table(os.path.join(RES_PATH, "Haplotype_Score_Info.tsv"), index_col="Haplotype_ID")
haplotype_score_info = haplotype_score_info.drop(columns=["Haplotype_Class"])
haplotype_score_info = haplotype_score_info.loc[total_best_haplotype_comb.index]

total_best_haplotype_comb = pd.concat([total_best_haplotype_comb, haplotype_score_info], axis=1, sort=False)
total_best_haplotype_comb.to_csv(os.path.join(RES_PATH, "Total_Best.Haplotypes_Combination.tsv"), sep="\t")

total_best_haplotype_comb

total_best_haplotype_comb["Hap_Score.by_Total_pop"].sum()
