==> 3-5-get-all-score.ipynb <==
import os, pickle
import numpy as np
import pandas as pd

RAW_PATH = "/data2/zhoujb/project/cowpea_project/rawData/"
SNP_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/snpDir/"
CLUSTER_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/SNPMMSeqCluster/"
RES_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/Result/"

genotype_hap_id = pd.read_table(os.path.join(RES_PATH, "Genotype_by_Hap_ID.tsv"), index_col="Region_index")

haplotype_score_info = pd.read_table(os.path.join(RES_PATH, "Haplotype_Score_Info.tsv"), index_col="Haplotype_ID")
haplotype_score_info = haplotype_score_info[['Hap_Score.by_Total_pop','Hap_Score.by_VL_pop', 'Hap_Score.by_G_pop', 'Hap_Score.by_VC_pop']].copy()

output_file = open(os.path.join(RES_PATH, "Samples_Score_Info.tsv"), "w")
print("\t".join(["Sample", "Class", "Mis_Hap_Num", 
                 "Hap_Score.by_Total_pop", "Hap_Score.by_VL_pop", "Hap_Score.by_G_pop", "Hap_Score.by_VC_pop"]), file=output_file)

detail_info = pd.read_excel(os.path.join(RAW_PATH, "Detail_information_344.xlsx"))
for index in detail_info.index:
    cultivar_id	= detail_info.loc[index, "Genotype No."]
    if cultivar_id not in genotype_hap_id.columns:
        continue
    cultivar_sub_class = detail_info.loc[index, "Types"]
    tmp_hap_id = [x for x in genotype_hap_id[cultivar_id].values if x != "-"]
    mis_hap_num = len([x for x in genotype_hap_id[cultivar_id].values if x == "-"])
    tmp_score = haplotype_score_info.loc[tmp_hap_id]
    tmp_sum = tmp_score.sum()
    final_score = [tmp_sum['Hap_Score.by_Total_pop']/854.768527747741*100, 
                   tmp_sum['Hap_Score.by_VL_pop']/952.5366599029439*100, 
                   tmp_sum['Hap_Score.by_G_pop']/547.8057376793577*100, 
                  tmp_sum['Hap_Score.by_VC_pop']/989.25449595043*100]
    
    print("\t".join([cultivar_id, cultivar_sub_class, str(mis_hap_num)]+[str(x) for x in final_score]), file=output_file)

output_file.close()

genotype_hap_id.loc["Chr09_35130001_35330000"]

genotype_hap_id = pd.read_table(os.path.join(RES_PATH, "Genotype_by_Hap_ID.tsv"), index_col="Region_index")

haplotype_score_info = pd.read_table(os.path.join(RES_PATH, "Haplotype_Score_Info.tsv"), index_col="Haplotype_ID")
haplotype_score_dict = haplotype_score_info["Hap_Score.by_Total_pop"].to_dict()

genotype_hap_score = pd.DataFrame(index=genotype_hap_id.index)
for cultivar_id in genotype_hap_id.columns[3:]:
    genotype_hap_score[cultivar_id] = genotype_hap_id[cultivar_id].map(haplotype_score_dict)

genotype_hap_score.index.name = ""
genotype_hap_score = genotype_hap_score.T
genotype_hap_score.to_csv(os.path.join(RES_PATH, "genotype_hap_score.txt"), sep="\t")
