==> 3-7-get-metabolome.ipynb <==
import os, pickle
import numpy as np
import pandas as pd

RAW_PATH = "/data2/zhoujb/project/cowpea_project/rawData/"
SNP_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/snpDir/"
CLUSTER_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/SNPMMSeqCluster/"
RES_PATH = "/data2/zhoujb/project/cowpea_project/basedXPXLR/Result/"

sample_index = pd.read_excel(os.path.join(RAW_PATH, "data_analysis_metabolome_20240112.xlsx"), sheet_name="样本信息")
sample_index = sample_index.dropna(subset=["MetID"])
sample_list = sample_index["MetID"].to_list()
index2sam_name = sample_index[["MetID", "Seq_ID"]].set_index("MetID").to_dict()["Seq_ID"]

deg_sample = pd.read_excel(os.path.join(RAW_PATH, "data_analysis_metabolome_20240112.xlsx"), sheet_name="deg_sample", index_col="ID")
rm_index = []
for index in deg_sample.index:
    change_0 = deg_sample.loc[index, 'Change   IF(LE2<1,"no",IF(LH2<0.67,"2L",IF(LH2>1.5,"1H","3nc")))']
    change_1 = deg_sample.loc[index, 'Change']
    change_2 = deg_sample.loc[index, 'Change.1']
    change_3 = deg_sample.loc[index, 'Change.2']
    if (change_0 == "no") and (change_1 == "no") and (change_2 == "no") and (change_3 == "no"):
        rm_index.append(index)

deg_sample = deg_sample.drop(index=rm_index)

deg_info = deg_sample[["KZ ID", "Compounds", "Class", "KEGG ID"]]
deg_info.to_csv(os.path.join(RAW_PATH, "metabolome_sample_info.txt"), sep="\t")

deg_sample_val = deg_sample[sample_list]
deg_sample_val = deg_sample_val.rename(columns=index2sam_name)
deg_val_mean = deg_sample_val.mean(axis=1)
deg_val_std = deg_sample_val.std(axis=1)
for col in deg_sample_val.columns:
    deg_sample_val[col] = (deg_sample_val[col] - deg_val_mean) / deg_val_std

deg_sample_val.index.name = ""
deg_sample_val = deg_sample_val.T

deg_sample_val.to_csv(os.path.join(RAW_PATH, "metabolome_sample_value.txt"), sep="\t")

deg_sample_val
